// src/pages/modules/pusat-karir/reference/specialization/sub-specialization/data/types.ts
export interface ISubSpecialization {
  id_spesialisasi: string
  id_sub_spesialisasi: string
  id_satuan_organisasi: string
  nama_spesialisasi: string
  nama_parent_spesialisasi: string
  slug: string
  urutan: number

  created_at: string
  created_user: string
  updated_at: string
  updated_user: string

  nama_user_created: string
  nama_user_updated: string
}